Clean up historical offload helpers
