+++ sei-db/state_db/ss/offload/consumer/scylla.go
@@ -134,21 +134,6 @@ func (s *scyllaSink) WriteBatch(ctx context.Context, records []Record) error {
 	return s.writeRecordsPipelined(ctx, records)
 }
 
-func compactRecords(records []Record) []Record {
-	for _, rec := range records {
-		if rec.Entry == nil {
-			out := make([]Record, 0, len(records))
-			for _, rec := range records {
-				if rec.Entry != nil {
-					out = append(out, rec)
-				}
-			}
-			return out
-		}
-	}
-	return records
-}
-
 func (s *scyllaSink) writeRecord(ctx context.Context, rec Record) error {
 	entry := rec.Entry
 	if entry == nil {
@@ -225,34 +210,6 @@ func (s *scyllaSink) writeRecordRows(ctx context.Context, version int64, entry *
 		})
 	}
 	return g.Wait()
-}
-
-type scyllaMutation struct {
-	storeName string
-	pair      *proto.KVPair
-}
-
-type scyllaMutationKey struct {
-	storeName string
-	key       string
-}
-
-func compactMutations(entry *proto.ChangelogEntry) []scyllaMutation {
-	mutations := make([]scyllaMutation, 0)
-	indexByKey := make(map[scyllaMutationKey]int)
-	for _, ncs := range entry.Changesets {
-		storeName := ncs.Name
-		for _, pair := range ncs.Changeset.Pairs {
-			key := scyllaMutationKey{storeName: storeName, key: string(pair.Key)}
-			if idx, ok := indexByKey[key]; ok {
-				mutations[idx].pair = pair
-				continue
-			}
-			indexByKey[key] = len(mutations)
-			mutations = append(mutations, scyllaMutation{storeName: storeName, pair: pair})
-		}
-	}
-	return mutations
 }
 
 func (s *scyllaSink) effectiveMutationWorkers() int {
